refactored auth pages folder structure
