+++ internal/tui/pages/auth/auth.go
@@ -1,4 +1,4 @@
-package pages
+package auth
 
 import (
 	"context"
